requirements.txt
discord.py==2.1.0
discord-py-slash-command==4.2.1
requests==2.28.1
beautifulsoup4==4.11.1
bs4==0.0.1